Refactor: CreateStudyForm Component로 분리
