--- src/components/organism/index.ts
@@ -1,2 +1,0 @@
-export * from './EditStudy/EditStudy';
-// export {};
